Moving tests to a specific html/js fil and fixing code style.
We are talking about functional tests here. Anyway, for the
future, if I move all the Anima logic, it would be good to
add also unit tests, with something "real" like qunit.

I also changed the style of the functions, from xx_yy to xxYy
the both tests and main code.