Showing only user-relevant columns
It was needed to change a little how getting the selected player was implemented,
as the index column is not at the DOM if it is hidden. It was needed to use
dataTables fnGetData for that.

I also added "Surprise" column, as we also want to show that. Computing the
string is something that is still needed.